% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acquire_mm22.R
\name{acquire_mm22}
\alias{acquire_mm22}
\title{Download a raw mm22.csv file}
\usage{
acquire_mm22(path)
}
\arguments{
\item{path}{Folder to put the downloaded data in. If missing a tempfile will
be created.}
}
\value{
Path and filename of the downloaded file.
}
\description{
\code{\link[=acquire_mm22]{acquire_mm22()}} is mainly a wrapper to download the file stored in
\code{\link[=mm22_url]{mm22_url()}}. You can optionally specify where to save the downloaded file.
}
\examples{
\dontrun{
file <- acquire_mm22()
data <- file |>
          get_mm22_metadata()

my_mm22 <- acquire_mm22("~/data")
}
}
