% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acquire_mm23.R
\name{acquire_mm23}
\alias{acquire_mm23}
\title{Download a raw mm23.csv file}
\usage{
acquire_mm23(path)
}
\arguments{
\item{path}{Folder to put the downloaded data in. If missing a tempfile will
be created.}
}
\value{
Path and filename of the downloaded file.
}
\description{
\code{\link[=acquire_mm23]{acquire_mm23()}} is mainly a wrapper to download the file stored in
\code{\link[=mm23_url]{mm23_url()}}. You can optionally specify where to save the downloaded file.
}
\examples{
\dontrun{
file <- acquire_mm23()
data <- file |>
          get_mm23_metadata()

my_mm23 <- acquire_mm23("~/data")
}
}
