% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acquire_ppi.R
\name{acquire_ppi}
\alias{acquire_ppi}
\title{Download a raw ppi.csv file}
\usage{
acquire_ppi(path)
}
\arguments{
\item{path}{Folder to put the downloaded data in. If missing a tempfile will
be created.}
}
\value{
Path and filename of the downloaded file.
}
\description{
\code{\link[=acquire_ppi]{acquire_ppi()}} is mainly a wrapper to download the file stored in
\code{\link[=ppi_url]{ppi_url()}}. You can optionally specify where to save the downloaded file.
}
\examples{
\dontrun{
file <- acquire_ppi()
data <- file |>
          get_ppi_metadata()

my_ppi <- acquire_ppi("~/data")
}
}
