% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acquire_prices.R
\name{acquire_prices}
\alias{acquire_prices}
\title{Download ONS shopping price dataset}
\usage{
acquire_prices(path)
}
\arguments{
\item{path}{Folder to put the downloaded data in. If missing a tempfile will
be created.}
}
\value{
The file path and name of the downloaded file.
}
\description{
\code{\link[=acquire_prices]{acquire_prices()}} is mainly a wrapper to download the file stored in
\code{\link[=prices_url]{prices_url()}}. You can optionally specify where to save the downloaded file.
}
\examples{
\dontrun{
file <- acquire_prices()

my_prices <- acquire_prices("~/data")
}
}
