% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpi_quote_archive.R
\name{archive_urls}
\alias{archive_urls}
\title{Dataset of archive price quote urls}
\usage{
archive_urls()
}
\value{
A dataframe of years and url stubs
}
\description{
Contains the url endings for all the price quote downloads for 2010-2019.
Each url stub needs to be appended to the base
"https://www.ons.gov.uk/file?uri=/economy/inflationandpriceindices/datasets/consumerpriceindicescpiandretailpricesindexrpiitemindicesandpricequotes/"
}
\examples{
archive_urls()
}
