% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contribution.R
\name{contribution}
\alias{contribution}
\title{Calculate contribution of a series to 'all items' 12 month rate}
\usage{
contribution(
  year,
  month,
  all_items_index,
  all_items_weight,
  component_index,
  component_weight
)
}
\arguments{
\item{component_weight}{}
}
\description{
Calculate contribution of a series to 'all items' 12 month rate
}
