% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{cpi_price_quotes_url}
\alias{cpi_price_quotes_url}
\title{URL for the page containing CPI price quote downloads}
\usage{
cpi_price_quotes_url()
}
\value{
url of the ONS price quote page
}
\description{
URL for the page containing CPI price quote downloads
}
\examples{
url <- cpi_price_quotes_url()
}
