% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpi_quote_archive.R
\name{get_cpi_price_quotes_archive}
\alias{get_cpi_price_quotes_archive}
\title{Get old price quotes from 2010-2019}
\usage{
get_cpi_price_quotes_archive(year, foodonly = TRUE)
}
\arguments{
\item{year}{the year(s) to get}

\item{foodonly}{whether to exclude non-food products from the output.}
}
\value{
a dataframe of price quotes
}
\description{
Get old price quotes from 2010-2019
}
\examples{
\dontrun{
get_cpi_price_quotes_archive(2015:2019)
}
}
