% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{get_cpi_weights_jan}
\alias{get_cpi_weights_jan}
\title{Get January CPI weights}
\usage{
get_cpi_weights_jan()
}
\value{
A dataframe of the January weights as published in Table 11 of the
CPI reference tables spreadsheet. They are not included in mm23 which only
has the weight at annual level.
}
\description{
Get January CPI weights
}
\examples{
\dontrun{
get_cpi_weights_jan()
}
}
