% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{get_cpih_cdid_lookup}
\alias{get_cpih_cdid_lookup}
\title{Get a reference table of CPIH CDIDs}
\usage{
get_cpih_cdid_lookup()
}
\value{
A data frame of CDIDs for overall index, weight, 1 month rate and 12
month rate for all CPIH series. Can be useful as a reference table.
}
\description{
Get a reference table of CPIH CDIDs
}
