% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mm23_metadata.R
\name{get_mm23_metadata}
\alias{get_mm23_metadata}
\title{Creates mm23 metadata}
\usage{
get_mm23_metadata(rawfile)
}
\arguments{
\item{rawfile}{A downloaded mm23.csv file. If missing will attempt to
download from ONS.}
}
\value{
A tibble containing mm23 metadata
}
\description{
Creates mm23 metadata
}
\examples{
\dontrun{
url <- "https://www.ons.gov.uk/file?uri=/economy/inflationandpriceindices/
datasets/consumerpriceindices/current/mm23.csv"
    tmp <- tempfile()
download.file(url, tmp)
rawfile <- tmp
get_mm23_month(rawfile)
}
}
