% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ppi_metadata.R
\name{get_ppi_metadata}
\alias{get_ppi_metadata}
\title{Creates ppi metadata}
\usage{
get_ppi_metadata(rawfile)
}
\arguments{
\item{rawfile}{A downloaded ppi.csv file. If missing will attempt to
download from ONS.}
}
\value{
A tibble containing ppi metadata
}
\description{
Creates ppi metadata
}
\examples{
\dontrun{
url <- "https://www.ons.gov.uk/file?uri=/economy/inflationandpriceindices/
datasets/consumerpriceindices/current/ppi.csv"
    tmp <- tempfile()
download.file(url, tmp)
rawfile <- tmp
get_ppi_metadata(rawfile)
}
}
