% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mm23_month.R
\name{get_ppi_month}
\alias{get_ppi_month}
\title{Get monthly PPI data}
\usage{
get_ppi_month(rawfile)
}
\arguments{
\item{rawfile}{a downloaded ppi.csv file. If missing will attempt to
download from ONS.}
}
\value{
A tibble containint date, CDID and value.
}
\description{
Get monthly PPI data
}
\examples{
\dontrun{
get_ppi_month()
}
}
