% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mm23_quarter.R
\name{get_ppi_quarter}
\alias{get_ppi_quarter}
\title{Get quarterly ppi data}
\usage{
get_ppi_quarter(rawfile)
}
\arguments{
\item{rawfile}{A downloaded ppi.csv file. If missing will attempt to
download from ONS.}
}
\value{
A tibble containing date, CDID and value.
}
\description{
Get quarterly ppi data
}
\examples{
\dontrun{
get_ppi_quarter()
}
}
