% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mm23_year.R
\name{get_ppi_year}
\alias{get_ppi_year}
\title{Get yearly ppi data}
\usage{
get_ppi_year(rawfile)
}
\arguments{
\item{rawfile}{A downloaded ppi.csv file. If missing will attempt to
download from ONS.}
}
\value{
A tibble containing date, CDID and value.
}
\description{
Get yearly ppi data
}
\examples{
\dontrun{
get_ppi_year()
}
}
