% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prices.R
\name{get_price_data}
\alias{get_price_data}
\title{Get average price data from the ONS Shopping Price tool}
\usage{
get_price_data(file)
}
\arguments{
\item{file}{A spreadsheet download of price data}
}
\value{
A tidy dataset of average prices
}
\description{
Get average price data from the ONS Shopping Price tool
}
\examples{
\dontrun{
data <- get_price_data()
}
}
