% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prices.R
\name{get_price_metadata}
\alias{get_price_metadata}
\title{Get average price metadata from the ONS Shopping Price tool}
\usage{
get_price_metadata(file)
}
\arguments{
\item{file}{A spreadsheet download of price data}
}
\value{
A dataset of price metadata
}
\description{
Get average price metadata from the ONS Shopping Price tool
}
\examples{
\dontrun{
metadata <- get_price_metadata()
}
}
