% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{get_weights}
\alias{get_weights}
\title{Create a weights dataset}
\usage{
get_weights(rawfile, measure = "cpi")
}
\arguments{
\item{rawfile}{A downloaded mm23.csv file. If missing will attempt to
download from ONS.}

\item{measure}{The inflation figure to get weights for. Defaults to "CPI",
can be "CPIH"}
}
\value{
A dataframe of monthly weights, including the January weights which
are not exposed in mm23 but only available in the reference tables
spreadsheet.
}
\description{
Create a weights dataset
}
\examples{
\dontrun{
weights <- get_weights()
cpih_weights <- get_weights(measure = "cpih")
}
}
