% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contribution.R
\name{I_x_Dec}
\alias{I_x_Dec}
\title{Calculate Ix(Dec) for an index}
\usage{
I_x_Dec(month, value)
}
\arguments{
\item{month}{A month number}

\item{value}{An index value}
}
\value{
An index value for December
}
\description{
An index for December based on previous January = 100
}
\examples{
\dontrun{
I_x_Dec(month, value)
}
}
