% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contribution.R
\name{I_x_Jan}
\alias{I_x_Jan}
\title{Calculate Ia(Jan) for an index}
\usage{
I_x_Jan(month, value)
}
\arguments{
\item{month}{A month number}

\item{value}{An index value}
}
\description{
Calculate Ia(Jan) for an index
}
