% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{ignore_unused_imports}
\alias{ignore_unused_imports}
\title{To get rid of NOTE abouot use of ggplot in vignette}
\usage{
ignore_unused_imports()
}
\description{
https://r-pkgs.org/dependencies-in-practice.html#how-to-not-use-a-package-in-imports
}
