% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{mm22_url}
\alias{mm22_url}
\title{Default url for the current mm22 PPI dataset}
\usage{
mm22_url()
}
\value{
url of mm22 data
}
\description{
Default url for the current mm22 PPI dataset
}
\examples{
url <- mm22_url()
}
