% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{mm23_url}
\alias{mm23_url}
\title{Default url for the current mm23 dataset}
\usage{
mm23_url()
}
\value{
url of mme23 data
}
\description{
Default url for the current mm23 dataset
}
\examples{
url <- mm23_url()
}
