% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{prices_url}
\alias{prices_url}
\title{Default url for the CPI average prices spreadsheet}
\usage{
prices_url()
}
\value{
url of data download spreadsheet
}
\description{
Its not clear yet if this url is stable, need to test over future months.
}
\examples{
url <- prices_url()
}
