% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{reftables_url}
\alias{reftables_url}
\title{Default url for the CPI reference tables spreadsheet}
\usage{
reftables_url()
}
\value{
url of reference tables
}
\description{
Its not clear yet if this url is stable, need to test over future months.
}
\examples{
url <- reftables_url()
}
