% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contribution.R
\name{contribution}
\alias{contribution}
\title{Calculate contribution of a series to 'all items' 12 month rate}
\usage{
contribution(
  year,
  month,
  all_items_index,
  all_items_weight,
  component_index,
  component_weight
)
}
\arguments{
\item{year}{A year number}

\item{month}{A month number}

\item{all_items_index}{All items (or category) index}

\item{all_items_weight}{All items (or category) weight}

\item{component_index}{Component item index}

\item{component_weight}{Component item weight}
}
\value{
A contribution value
}
\description{
Calculate contribution of a series to 'all items' 12 month rate
}
