% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contribution.R
\name{contribution}
\alias{contribution}
\title{Calculate contribution of a series to 'all items' 12 month rate}
\usage{
contribution(
  month,
  all_items_index,
  all_items_weight,
  component_index,
  component_weight
)
}
\arguments{
\item{month}{A month number}

\item{all_items_index}{All items (or category) index}

\item{all_items_weight}{All items (or category) weight}

\item{component_index}{Component item index}

\item{component_weight}{Component item weight}
}
\value{
A contribution value
}
\description{
The contribution function implements the formula for calculating the
contribution of a cpi component to its parent level, as set out in the ONS
\href{https://www.ons.gov.uk/economy/inflationandpriceindices/methodologies/consumerpricesindicestechnicalmanual2019#weights}{CPI Technical Manual}.

The function must be applied to an unchained index. The \code{\link[=unchain]{unchain()}} function
can be used to unchain a series. The formula is valid for data after 2018
when the CPI weights updated twice a year, and is set out below.

\eqn{c = component\ c}

\eqn{a =\ 'all\ items'\ CPI\ index}

\eqn{W^c_{y} = weight\ of\ component\ c\ in\ year\ y}

\eqn{I^c_t = index\ for\ component\ c\ in\ month\ t\ based\ on\ January\ of\ current\ year =100}

\eqn{I^a_{Jan} = all\ items\ index\ for\ January\ based\ on\ previous\ month\ (December) = 100}

\eqn{I^a_{Dec} = all\ items\ index\ for\ December\ based\ on\ previous\ January = 100}

\deqn{
(\frac{W^c_{y-1}} {W^a_{y-1}})
\times
(\frac{({I^c_{Dec}} - {I^c_{t-12}})}{I^a_{t-12}})
\times
100
\ \ +\ \
(\frac{W^c_y}{W^a_y})
\times
(\frac{({I^c_{Jan}}-100)}{I^a_{t-12}})
\times
I^a_{Dec}
\ \ +\ \
(\frac{W^c_y}{W^a_y})
\times
(\frac{(I^c_t - 100)}{I^a_{t-12}})
\times
\frac{I^a_{Jan}}{100}
\times
I^a_{Dec}
}{asciitext}

You should apply \code{\link[=contribution]{contribution()}} to a wide dataframe: one that has a field
containing the dates for the series, and columns containing the relevant
series indices and their weights.
}
\examples{
\dontrun{
bread = contribution(month = month(date),
  all_items_index = L523,
  all_items_weight = L5CZ,
  component_index = L52I,
  component_weight = L5DH)
}
}
