% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cpi_quotes.R
\name{get_cpi_price_quotes}
\alias{get_cpi_price_quotes}
\title{Get CPI price quotes for food items.}
\usage{
get_cpi_price_quotes(year, month, path)
}
\arguments{
\item{year}{Year to extract}

\item{month}{Month to extract}

\item{path}{Optional path to save data}
}
\value{
a tibble of price quotes
}
\description{
Get CPI price quotes for food items.
}
\examples{
\dontrun{
get_cpi_price_quotes(2021, 8)
}
}
