% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contribution.R
\name{I_x_Dec}
\alias{I_x_Dec}
\title{Calculate Ia(Dec) for an index}
\usage{
I_x_Dec(month, value)
}
\arguments{
\item{month}{A month number}

\item{value}{An index value}
}
\description{
Calculate Ia(Dec) for an index
}
