% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpi_quote_archive.R
\name{make_cpi_quote_archive}
\alias{make_cpi_quote_archive}
\title{Create an rds containing 2010-2019 CPI price quotes.}
\usage{
make_cpi_quote_archive(dir = getwd())
}
\arguments{
\item{dir}{The folder to put the final archive into.}
}
\value{
creates an rds file in the chosen folder
}
\description{
Create an rds containing 2010-2019 CPI price quotes.
}
\examples{
\dontrun{
make_cpi_quote_archive(dir = "~")
}
}
