% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{mm23_url}
\alias{mm23_url}
\title{Stores the url of the latest mm23.csv data}
\usage{
mm23_url()
}
\value{
a url string
}
\description{
Stores the url of the latest mm23.csv data
}
\examples{
\dontrun{
mm23_url()
}
}
