% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{unchain}
\alias{unchain}
\title{Hopefully unchains a CPI series}
\usage{
unchain(month, value)
}
\arguments{
\item{month}{The month number for the index}

\item{value}{The index value}
}
\value{
An unchained index value
}
\description{
Requires a data set with a month variable, uses dplyr::lag to unchain using
previous values. Assumes that CPI weights change twice a year.
}
